Add files via upload
Issue#1: the 'shapely' package mentioned in the requirements.txt is not compatible with the notebook format for pickle. Hence it's throwing the "Multipolygon is not an iterable error.

Fix: Downgrade the shapely pkg version 

Issue#2: keyerror: 'r_lc_name'

Fix: this is error in default script for coco train/val split. 
the coco dataframe doesn't contain any column named 'r_lc_name' but 'lc_name' in the original package